### 2024-06-11 #6872 @editorialbot
[PRE REVIEW]: Back to sequences: find the origin of k-mers
**Submitting author:** @pierrepeterlongo (<a href="http://orcid.org/0000-0003-0776-6407">Pierre Peterlongo</a>)
**Repository:** https://github.com/pierrepeterlongo/back_to_sequences/
**Branch with paper.md** (empty if default branch):
**Version:** v0.6.5
