+++ paper/paper.bib
@@ -30,6 +30,7 @@
   number={1},
   pages={209},
   year={2022},
+doi={10.1186/s13059-022-02771-2},
   publisher={Springer}
 }
 
@@ -42,6 +43,7 @@
   number={Suppl 11},
   pages={S6--S12},
   year={2009},
+doi=‘10.1038/nmeth.1376},
   publisher={Nature Publishing Group US New York}
 }
 
@@ -53,6 +55,7 @@
   number={9},
   pages={1165--1173},
   year={2010},
+doi={10.1101/gr.101360.109},
   publisher={Cold Spring Harbor Lab}
 }
 @article{benoit2016multiple,
@@ -62,6 +65,7 @@
   volume={2},
   pages={e94},
   year={2016},
+doi={10.7717/peerj-cs.94},
   publisher={PeerJ Inc.}
 }
 @article{li2012comparison,
@@ -72,6 +76,7 @@
   number={1},
   pages={25--37},
   year={2012},
+doi={10.1093/bfgp/elr035},
   publisher={Oxford University Press}
 }
 
@@ -82,6 +87,7 @@
   volume={20},
   pages={1--20},
   year={2019},
+doi={10.1186/s13059},
   publisher={Springer}
 }
 
@@ -93,6 +99,7 @@
   number={12},
   pages={i283--i292},
   year={2014},
+doi={10.1093/bioinformatics/btu288},
   publisher={Oxford University Press}
 }
 
@@ -105,6 +112,7 @@
   number={2},
   pages={e11--e11},
   year={2015},
+doi={10.1093/nar/gku1187},
   publisher={Oxford University Press}
 }
 
@@ -116,6 +124,7 @@
   number={1},
   pages={vbac029},
   year={2022},
+doi={10.1093/bioadv/vbac029},
   publisher={Oxford University Press}
 }
 
@@ -126,6 +135,7 @@
   journal={Genome Research},
   pages={gr--277615},
   year={2023},
+doi={10.1101/gr.277615.122},
   publisher={Cold Spring Harbor Lab}
 }
 
@@ -138,6 +148,7 @@
   number={5},
   pages={603--615},
   year={2002},
+doi={10.1145/585265.585267},
   publisher={ACM New York, NY, USA}
 }
 
@@ -150,6 +161,7 @@
   number={5},
   pages={455--477},
   year={2012},
+doi={10.1089/cmb.2012.0021},
   publisher={Mary Ann Liebert, Inc. 140 Huguenot Street, 3rd Floor New Rochelle, NY 10801 USA}
 }
 
@@ -160,6 +172,7 @@
   volume={274},
   pages={92--102},
   year={2020},
+doi={10.1016/j.dam.2018.03.035},
   publisher={Elsevier}
 }
 @article{wood2019improved,
@@ -169,6 +182,7 @@
   volume={20},
   pages={1--13},
   year={2019},
+doi={10.1186/s13059-019-1891-0},
   publisher={Springer}
 }
 
@@ -188,13 +202,14 @@
 
 %kmdiff
 @article{lemane2022k,
-  title={k mdiff, large-scale and user-friendly differential k-mer analyses},
+  title={kmdiff, large-scale and user-friendly differential k-mer analyses},
   author={Lemane, T{\'e}o and Chikhi, Rayan and Peterlongo, Pierre},
   journal={Bioinformatics},
   volume={38},
   number={24},
   pages={5443--5445},
   year={2022},
+doi={10.1093/bioinformatics/btac689},
   publisher={Oxford University Press}
 }
 
@@ -207,6 +222,7 @@
   volume={16},
   pages={1--10},
   year={2015},
[…]
