### 2024-06-11 f2a8d42
Update paper.bib - typo
