+++ paper/paper.bib
@@ -43,7 +43,7 @@ doi={10.1186/s13059-022-02771-2},
   number={Suppl 11},
   pages={S6--S12},
   year={2009},
-doi=‘10.1038/nmeth.1376},
+doi={10.1038/nmeth.1376},
   publisher={Nature Publishing Group US New York}
 }
 
