### 2024-06-11 847bf27
Update paper.bib - corrected a doi
