+++ paper/paper.bib
@@ -87,7 +87,7 @@ doi={10.1093/bfgp/elr035},
   volume={20},
   pages={1--20},
   year={2019},
-doi={10.1186/s13059},
+doi={10.1186/s13059-019-1632-4},
   publisher={Springer}
 }
 
