### 2024-06-19 #6872 @pierrepeterlongo
> **Submitting author:** @pierrepeterlongo ([Pierre Peterlongo](http://orcid.org/0000-0003-0776-6407)) **Repository:** https://github.com/pierrepeterlongo/back_to_sequences/ **Branch with paper.md** (empty if default branch): **Version:** v0.6.5 **Editor:** Pending **Reviewers:** Pending **Managing EiC:** Kevin M. Moerman
>
> ## Status
> [![status](https://camo.githubusercontent.com/9b49e5511eee7ce37db8c8817ada82c8b6ff44065eae0263e8e19fce6382a9cb/68747470733a2f2f6a6f73732e7468656f6a2e6f72672f7061706572732f32623230643066613238373130396662306663646233396234386238316232312f7374617475732e737667)](https://joss.theoj.org/papers/2b20d0fa287109fb0fcdb39b48b81b21)
>
> Status badge code:
>
> ```
> ```
>
> **Author instructions**
>
> Thanks for submitting your paper to JOSS @pierrepeterlongo. **Currently, there isn't a JOSS editor assigned** to your paper.
>
> @pierrepeterlongo if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
>
> **Editor instructions**
>
> The JOSS submission bot @editorialbot is here to help you find and assign reviewers and start the main review. To find out what @editorialbot can do for you type:
>
> ```
> @editorialbot commands
> ```

Suggested reviewers:
[Brice Letcher](https://reviewers.joss.theoj.org/reviewers/1804)
[Anuradha Wickramarachchi](https://reviewers.joss.theoj.org/reviewers/3975)
[Michael B. Hall](https://reviewers.joss.theoj.org/reviewers/2619)
