### 2024-06-24 f31cbe8
Update README.md - added the JOSS proof PDF
