### 2024-06-26 2f91a2f
v0.6.6 - corrected a bug when using multiple input files
