+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "back_to_sequences"
-version = "0.6.5"
+version = "0.6.6"
 edition = "2021"
 
 authors = [
