### 2024-07-03 #6872 @Kevin-Mattheus-Moerman
@pierrepeterlongo thanks for making those updates/changes to have your paper conform more to our standards. As AEiC for this track I am currently pre-reviewing your submission and am also looking for a handling editor. Since all editors in this domain are pre-occupied with other submissions it may take a longer than usual to find/assign an editor. Hence I've just added the `waitlisted` label. We'll assign an editor as soon as one becomes available.

For the moment could you work on the below points:

- [ ] I have only had a quick look at your submission so far, but the README does not seem to link to: 1) automated testing, and 2) dedicated project documentation. Please add these and link to them clearly in the README.
