+++ README.md
@@ -5,16 +5,6 @@
 
 <img src="k2s.jpg" alt="Old library - IA generated" width="150" height="150">
 
-<!-- Add a table of content -->
-* [Description](#Description)
-* [Citation](#Citation)
-* [Install](#Install)
-* [Quick benchmark (obtained with version v0.6.4)](#Quickbenchmarkobtainedwithversionv0.6.4)
-* [Usage](#Usage)
-* [Generate random data for testing](#Generaterandomdatafortesting)
-
-<!-- ![Old library - IA generated](k2s.jpg) -->
-##  1. <a name='Description'></a>Description
 
 Given a set $K$ of kmers (fasta / fastq [.gz] format) and a set of sequences  (fasta / fastq [.gz] format), this tool will extract the sequences containing some of those kmers.
 
@@ -25,239 +15,9 @@ TGGATAAAAAGGCTGACGAAAGGTCTAGCTAAAATTGTCAGGTGCTCTCAGATAAAGCAGTAAGCGAGTTGGTGTTCGCT
 ```
 In this case 20 kmers are shared with the indexed kmers. This represents 6.13% of the kmers in the sequence.
 
-##  2. <a name='Citation'></a>Citation
 
-> Anthony Baire, Pierre Peterlongo
-[Back to sequences: find the origin of kmers](https://doi.org/10.1101/2023.10.26.564040). bioRxiv 
-2023.10.26.564040; doi: https://doi.org/10.1101/2023.10.26.564040
-
-```bibtex
-@article{baire_back_2023,
-	title = {Back to sequences: find the origin of kmers},
-	url = {https://www.biorxiv.org/content/early/2023/10/29/2023.10.26.564040},
-	doi = {10.1101/2023.10.26.564040},
-	journaltitle = {{bioRxiv}},
-	author = {Baire, Anthony and Peterlongo, Pierre},
-	date = {2023},
-}
-```
-
-The paper is submitted to [JOSS](https://joss.theoj.org/), which generated the following [pdf](https://github.com/openjournals/joss-papers/blob/joss.06872/joss.06872/10.21105.joss.06872.pdf).
-
-##  3. <a name='Install'></a>Install
-
-```bash
-git clone https://github.com/pierrepeterlongo/back_to_sequences.git
-cd back_to_sequences
-RUSTFLAGS="-C target-cpu=native" cargo install --path .
-```
-
-A test can be performed by running `cd tiny_test; sh tiny_test.sh; cd -`.
- 
-##  4. <a name='Quickbenchmarkobtainedwithversionv0.6.4'></a>Quick benchmark (obtained with version v0.6.5)
-This benchmark is reproducible by running `generate_data.sh` and then `bench.sh` in the `benchs` folder. 
-Presented results were obtained on 
-* the GenOuest platform on a node with 32 threads Xeon 2.2 GHz, denoted by "genouest" in the table below.
-* a MacBook, Apple M2 pro, 16 GB RAM, with 10 threads, denoted by "mac" in the table below.
-* AMD Ryzen 7 4.2 GHz 5800X 64 GB RAM,  with 16 threads, denoted by "AMD" in the table below.
-
-We indexed: one million kmers (exactly 995,318) of length 31.
-
-We queried: from 10,000 reads to 200 million reads (+ 1 billion on the cluster), each of length 100.
-
-| Number of reads | Time genouest | Time mac |  Time AMD | max RAM |
-|-----------------|----------|---|---|---|
-| 10,000          | 0.7s  | 0.54s | 0.4s | 0.13 GB |
-| 100,000         | 0.8s  | 0.8s | 1.2s |  0.13 GB |
[…]
