+++ TODO.md
@@ -1,0 +1,11 @@
+
+
+# TODO
+* [X] Add a validation test (04/10/2023)
+* [X] Add a number of shared kmers per sequence instead of only their ratio 
+* [X] Add a threshold on the number of shared kmers
+* [X] Parallelize the read extraction step
+* [ ] Thinks about a way to adapt this to protein sequences
+* [X] Add an option to set the size of the bloom filter used by kmindex
+* [ ] Provide a way to index and query more than one set $K$ of kmers
+* [X] Output the strand of matched kmers
