### 2024-07-03 #6872 @pierrepeterlongo
Thanks @Kevin-Mattheus-Moerman for your feedbacks.

 > Please study the above reference check ☝️ and see if you can address any of the reported potential DOI issues. You can add/amend DOI entries in your .bib file, and call @editorialbot check references here to check them again.

The last entries with no DOI are citations of github repositories. Maybe this is not the right way to do?

> Your project appears to lack contributing guidelines at the moment. Please work to add these and to mention/link to them in your README. You could for instance create a CONTRIBUTING.md file (see here for some examples: https://contributing.md/example/).

We have added a contributing.md file indicating contributing guidelines.

> I have only had a quick look at your submission so far, but the README does not seem to link to: 1) automated testing, and 2) dedicated project documentation. Please add these and link to them clearly in the README.

The documentation is now in a separate readthedocs repository: https://b2s-doc.readthedocs.io/en
We will inform you when the link to automated testing is ready

Thanks,
Pierre
