### 2024-07-14 #6872 @pierrepeterlongo
Hello there,

Links to automated testing are now ready and linked in the readme.
All points you mentioned are now fixed.

Pierre
