### 2024-07-30 #6872 @Kevin-Mattheus-Moerman
@pierrepeterlongo I've invited an editor who I hope can handle this submission. I have looked at the automated testing link. Are you referring to the testing badge? That link does not seem to work. Could you also link to the contributing.md file from the readme?
