### 2024-07-31 #6872 @majensen
Hi @amoeba - could I get your help on [this work](#6872) for [JOSS](https://joss.theoj.org)? I think it's right up your alley. --@majensen
