### 2024-08-19 pierrepeterlongo/back_to_sequences#6 @Anjan-Purkayastha
Suggestions on improving your manuscript and documentation.
This is a `Good Enhancement Suggestion`.
I have reviewed your manuscript. While it describes a general-purpose bioinformatics tool that may have many applications. it will be helpful if you provide an overview of your k-mer search method and specifically highlight **how your tool solves the problem of preserving the k-mer/parent sequence link**.  Afer all, this is the central claim of your paper and I as a user of your tool am interested in how you have solved this problem. This should be included in both your manuscript as well as your online documentation.
The second suggestion is a minor one. You mention some code run-times in Table 1 of your manuscript. **Please be specific as to which time (real/user) you are reporting.**
