### 2024-08-19 #7066 @Anjan-Purkayastha
@majensen: Have completed installing and testing the tool. It is a useful addition to the set of general-purpose bioinformatics tools we use for processing NGS data.
I have left some good enhancement suggestion at this [link](https://github.com/pierrepeterlongo/back_to_sequences/issues/6#issue-2472472790), for the authors. Aside from this, they need to update their citations, and they should be good to go.
This is my first time interacting over GitHub, please do let me know if there is anything else I need to do to complete this review. Cheers.
