### 2024-08-20 1705fd0
Remove "Availability line"
I think this should be removed because a link to the software is given elsewhere in the generated paper. If the authors wish to keep it, I suggest using a full sentence, probably as the final sentence of the paragraph. Something like "The source code and documentation for back to sequences can be found at [pierrepeterlongo/back_to_sequences](https://github.com/pierrepeterlongo/back_to_sequences)."
