### 2024-08-20 f1d31e0
First usage of acronym requires definition
Feel free to swap out "memory" if you prefer but note you use RAM later on.
