+++ paper/paper.md
@@ -92,7 +92,7 @@ abundance of original reads, as the link between each original read and
 each of its $k$-mers would have to be stored. For instance, considering
 $k$-mers from a sequencing experiment of a human genome ($\approx 3$
 billion nucleotides) with a coverage of 50x (each $k$-mer occurs on
-average in 50 distinct reads) would require more than 2Tb of space
+average in 50 distinct reads) would require more than 2TB of space
 considering 64 bits for storing each link and 64 bits for storing the
 associated read identifier. This is not acceptable.
 
