### 2024-08-20 ef06b9a
Remove extra sections not allowed by JOSS
