+++ paper/paper.md
@@ -119,102 +119,7 @@ In this context, we propose `back_to_sequences`, a tool specifically
 dedicated to extracting from $\mathcal{S}$, a set of sequences (eg.
 reads), those that contain some of the $k$-mers from a set $\mathcal{K}$
 given as input. The occurrences of $k$-mers in each sequence of the
-queried set $\mathcal{S}$ can also be output. In addition to this
-fundamental task, additional features are described in the following.
-
-# Results
-
-The tool we introduce, `back_to_sequences`, uses the native `rust` data
-structures to index and query $k$-mers. Its advantages are 1. its
-simplicity (install and run); 2. its low RAM usage; 3. its fast running
-time; 4. its additional features adapted to $k$-mers: control of their
-orientation (see below) and sequence filtration based on the minimal and
-maximal percent of $k$-mers shared with the indexed set. To the best of
-our knowledge, there exists no other tool dedicated to this specific
-task.
-
-In the general case, the DNA sequence orientation is unknown. DNA can be
-sequenced either in one strand (eg $AAGGC$) or in the reverse complement
-strand, read from right to left and commutating $A$ with $T$, and $C$
-with $G$ (eg $GCCTT$). This is why $k$-mers from $\mathcal{S}$ and
-$\mathcal{K}$ can be considered either as original or as *canonical*,
-the smallest alphabetical word between the $k$-mer and its reverse
-complement.
-
-We propose a simple performance benchmark to assess the
-`back_to_sequences` scaling performances on random sequences and on real
-metagenomic data.
-
-## Benchmark on random sequences {#ssec:random_banch}
-
-We generated a random sequence $s$ on the alphabet $\{A,C,G,T\}$ of size
-100 million base pairs. From this sequence $s$, we randomly extracted
-50,000 sub-sequences each of size 50. We consider these sequences as
-containing the set $\mathcal{K}$ of $k$-mers to be searched. As we used
-$k=31$, each subsequence contains $50-31+1 = 20 kmers$. Doing so, we
-consider a set of at most $50000\times 20 = 1,000,000$ $k$-mers.
-
-We also generated six sets:
-$\{\mathcal{S}_{10k}, \mathcal{S}_{100k}, \mathcal{S}_{1M}, \mathcal{S}_{10M}, \mathcal{S}_{100M},  \mathcal{S}_{200M}\}$,
-composed respectively of 10 thousand, 100 thousand, one million, 10
-million, 100 million, and 200 million sequences, each of length 100
-nucleotides. Each sequence is randomly sampled from $s$.
-
-In each sequence of each set $\mathcal{S}_i$, we searched for the
-existence of $k$-mers indexed in $\mathcal{K}$. The performances are
-provided Table [1](#tab:res_bench){reference-type="ref"
-reference="tab:res_bench"}. Presented results were obtained on the
-GenOuest platform on a node with 32 threads Xeon 2.2 GHz, on a
-MacBook, Apple M2 pro, 16 GB RAM with 10 threads, and an AMD Ryzen
-7 4.2 GHz 5800X 64 GB RAM with 16 threads, respectively denoted by "Time GenOuest",
- "Time mac" and "Time AMD" in this table. These results highlight
-the scalability of the `back_to_sequences` tool, able to search a
-million of $k$-mers in hundreds of millions of reads on a laptop
-computer in a matter of dozens of minutes with negligible RAM usage.
-
-::: {#tab:res_bench}
-    Number of reads  Time GenOuest   Time mac   Time AMD   max RAM
-  ----------------- --------------- ---------- ---------- ---------
-             10,000      0.7s          0.5s       0.4s     0.13 GB
-            100,000      0.8s          0.8s       1.2s     0.13 GB
[…]
