### 2024-08-20 b8904f4
Add missing References section
