+++ paper/paper.md
@@ -8,13 +8,13 @@ tags:
 authors:
   - name: Anthony Baire
     affiliation: 1
-  - name: Pierre Marijon 
+  - name: Pierre Marijon
     orcid: 0000-0002-6694-6873
     affiliation: 2
   - name: Francesco Andreace
     orcid: 0009-0008-0566-200X
     affiliation: 3
-  - name: Pierre Peterlongo 
+  - name: Pierre Peterlongo
     orcid: 0000-0003-0776-6407
     affiliation: 1
 affiliations:
@@ -25,7 +25,7 @@ affiliations:
  - name: Department of Computational Biology, Institut Pasteur, Université Paris Cité, Paris, F-75015, France
    index: 3
 date: 10 June 2024
-bibliography: paper.bib 
+bibliography: paper.bib
 
 
 ---
