### 2024-08-20 pierrepeterlongo/back_to_sequences#8 @amoeba
JOSS Review
Ref my checklist: https://github.com/openjournals/joss-reviews/issues/7066#issuecomment-2297831057

Hi @pierrepeterlongo, I've had a chance to sit down with https://github.com/pierrepeterlongo/back_to_sequences/ and things look mostly in order. This looks like a generally useful piece of software and I appreciate the effort you've put into the source code, documentation, and the paper.

Below are my review comments, all of which should be addressed before acceptance.

## Functionality

_Needs changes_

  - I wasn't able to run the benchmarks as-is. It looks like the data generation script and the benchmarking script don't agree about what size reads to benchmark. I edited the file manually to make the benchmarks run and was able to confirm the software performs about as fast as you say. Please fix the benchmark scripts so they run out of the box.

## Documentation

_Needs changes_

- Because the JOSS submission is an archive of the source code, it's important that the documentation for the software is contained within that archive. While [the documentation
](https://b2s-doc.readthedocs.io/) is good, I noticed that it lives in a repo separate from the source code being submitted here. Therefore, I suggest the documentation in https://github.com/pierrepeterlongo/b2s_doc be moved into https://github.com/pierrepeterlongo/back_to_sequences/. Let me know if this isn't possible for some reason and we can probably work something out.
  - This is missing. Please add it in the README and also in the docs website.
- Overall the main README is a bit sparse. It would be good if it had sections for Installation, Usage (it does now), Benchmarks, and Community Guidelines. It could still link to the full documentation website.

## Software paper

_Needs changes_

The paper was generally well-written but contained too much information for JOSS. See https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain for the allowed sections. [I've filed a PR](https://github.com/pierrepeterlongo/back_to_sequences/pull/7) to remove the sections which also includes a variety of other fix-ups. The content I removed is very interesting and I think it would be good to retain it in some form. Maybe on the documentation website?

Please let me know if you have any questions. Thanks for your time.
