### 2024-08-20 #7066 @amoeba
Hi @majensen and @pierrepeterlongo, I've finished my review and [filed an issue on your repo](https://github.com/pierrepeterlongo/back_to_sequences/issues/8). I do ask for some changes there so please have a look and let me know when you'd like me to re-review.
