### 2024-08-25 #7066 @majensen
Hi @pierrepeterlongo - the reviewers have made their comments and have items to address at https://github.com/pierrepeterlongo/back_to_sequences/issues/6 and https://github.com/pierrepeterlongo/back_to_sequences/issues/8. Please keep us informed on your progress here as you work through these. Thanks!
