### 2024-08-26 pierrepeterlongo/back_to_sequences#6 @pierrepeterlongo
Hi @Anjan-Purkayastha
Thanks a lot for your review and these suggestions. This will improve the manuscript and documentation.
