### 2024-08-27 459d5e7
Add infos in the README
