### 2024-08-27 pierrepeterlongo/back_to_sequences#6 @pierrepeterlongo
Hello @Anjan-Purkayastha

> it will be helpful if you provide an overview of your k-mer search method and specifically highlight how your tool solves the problem of preserving the k-mer/parent sequence link.

This is now something that is proposed in the documentation: https://b2s-doc.readthedocs.io/en/latest/methods.html

> This should be included in both your manuscript as well as your online documentation.

This appears only in the documentation. This is due to the fact that
 - with respect to this [recommendation](https://github.com/openjournals/joss-reviews/issues/6872#issuecomment-2205642997), we decided to extract the documentation from the README of the main repo.
 - reviewer @amoeba [suggested](https://github.com/pierrepeterlongo/back_to_sequences/issues/8) halving the size of the manuscript in order to comply with JOSS recommendations. Hence most of the paper content is now located in the [documentation](https://b2s-doc.readthedocs.io)

> You mention some code run-times in Table 1 of your manuscript. Please be specific as to which time (real/user) you are reporting.

This is now clearly indicated in the [bench result page](https://b2s-doc.readthedocs.io/en/latest/results.html#id1).

Thanks again for your suggestions
Pierre
