### 2024-08-27 pierrepeterlongo/back_to_sequences#8 @pierrepeterlongo
Hello @amoeba

> I wasn't able to run the benchmarks as-is

This is now fixed. And a new cleaning script was added
