### 2024-08-27 pierrepeterlongo/back_to_sequences#8 @pierrepeterlongo
> I suggest the documentation in https://github.com/pierrepeterlongo/b2s_doc be moved into https://github.com/pierrepeterlongo/back_to_sequences/.

We have made the decision to separate the documentation from the github repo because the last item of this [recommendation](https://github.com/openjournals/joss-reviews/issues/6872#issuecomment-2205642997). It seems to me that this separation is required, even if I'm not 100% sure of this claim.

Anyway we added specific pieces of information in the README.
