### 2024-08-27 pierrepeterlongo/back_to_sequences#8 @pierrepeterlongo
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

We now clearly link the [CONTRIBUTING](https://github.com/pierrepeterlongo/back_to_sequences/blob/main/CONTRIBUTING.md) file in the README.
