### 2024-08-27 pierrepeterlongo/back_to_sequences#8 @pierrepeterlongo
> Overall the main README is a bit sparse. It would be good if it had sections for Installation, Usage (it does now), Benchmarks, and Community Guidelines. It could still link to the full documentation website.

As mentioned we updated the README. We can further change the balance between this README and the external documentation if required.
