### 2024-08-27 pierrepeterlongo/back_to_sequences#8 @pierrepeterlongo
> remove the sections which also includes a variety of other fix-ups.

Thank you again for this important work.
We have applied the pull request and adjusted the rest of the manuscript. The parts you removed have been transferred to the documentation.
