+++ .github/workflows/lints.yaml
@@ -22,7 +22,7 @@ jobs:
           toolchain: ${{ matrix.toolchain }}
           components: rustfmt
       - name: Run fmt
-        run: cargo fmt --all --check
+        run: cargo fmt --all
 
   clippy:
     strategy:
