### 2024-08-27 #7066 @pierrepeterlongo
Hello @majensen,

We took in consideration reviewer's comments. I closed the related issues. https://github.com/pierrepeterlongo/back_to_sequences/issues/6 and https://github.com/pierrepeterlongo/back_to_sequences/issues/8.

Best,
Pierrre
