### 2024-09-04 #7066 @amoeba
Hi @majensen: Following the submitting author accepting all of my changes in https://github.com/pierrepeterlongo/back_to_sequences/issues/8, I have checked off the remaining items in [my checklist](https://github.com/openjournals/joss-reviews/issues/7066#issuecomment-2297831057).

The final issue I have is that the dedicated documentation for the project is kept in a [separate GitHub repo](https://github.com/pierrepeterlongo/b2s_doc) and my worry is that it won't be archived alongside the source code. I think it's important for the documentation to be archived by JOSS so I wonder if I wonder if JOSS doesn't care, whether JOSS can archive both repositories, or if I should push the author to integrate the two repositories. Please advise @majensen.
