### 2024-09-05 #7066 @majensen
@amoeba - I think this is a worthy consideration. I don't think we need to ask for a change in the code organization. What I see on Zenodo is a place for "Related" artifacts with associated metadata:

So my recommendation would be, create an additional Zenodo archive with the docs repo, and then add the DOI in the "Related" section, with the tag "is documented by". Thoughts?
