### 2024-09-05 #7066 @amoeba
I think that would be sufficient. Thanks for coming up with a solution. My review is now an Accept.
