### 2024-09-05 #7066 @pierrepeterlongo
Hey,

Thanks for this suggestion.
I'd like to be sure to understand your suggestion.

You suggest a Zenodo archive for the doc, thus generating a DOI, say DOI_doc
+
a Zenodo archive for the code, indicating the DOI_doc as a "is documented by" relation

Do I understand you correctly?
