### 2024-09-06 #7066 @majensen
Exactly right @pierrepeterlongo - as one of the last steps (after @Anjan-Purkayastha does his final check), I'll ask you to create a Zenodo archive of the code and send us back the DOI. In this case, we'd ask you to also create an archive for the docs repo and annotate the code repo as you describe. You can do this at any time, but may be worth waiting until we get the final check.
