### 2024-09-16 #7066 @majensen
@pierrepeterlongo: can you consider making or responding to the following suggestions in the MS (line numbers of latest PDF in this thread):

Title: capitalize “Find”?

l.13: “finding back” -> “recovering”
l.16: “occurences” -> distribution?
l.17: “short read” -> “short reads”

l.20 - may be more readable with each step in its own paragraph (like a bulleted list); I would capitalize the first word in each step.

l.37: “it has been…” -> “this has been…”
l.41: “only a few.” -> “only a few examples.”

l.55: “associated to” -> “associated with”
l.56: delete “to cite a few”

l.59 “Finding back” -> “Recovering”
l.60 “hardly scales” -> “is too costly for”
l.62 “even if they were” -> “even though they have been”
l.65 “but not extract” -> “but do not extract”

l.67 “eg.” -. “e.g.,”
l.75 “the feature to extract” -> “a feature for extracting”
l.79 “one k-mer” -> “one k-mer (length 31)” (correct?)

l.80: link “https://b2s-doc.readthedocs.io/en/latest/benchmark.html” is broken;
https://b2s-doc.readthedocs.io/en/latest/results.html looks promising, but doesn’t contain the results for the alternative apps.

l.87 “Summing up, these…” -> “Summing up, we find these…”
l.87 “not meant for” -> “not appropriate for”

l.95 “finding back” -> “recovering”
l.97-98 “applications, such as” -> “applications, in such areas as”
l.99. “datasets, all of which”. I would make a separate final sentence, like “Because of the efficiency of our approach, such applications could be executed in real time during the sequencing process.”

Refs

l.113: “de bruin” -> “de Bruin”
l.118: “bloom filters” -> “Bloom filters”
l.132: “bloom filters” -> “Bloom filters”
l.156: “kraken 2” -> “Kraken 2”
