+++ src/lib.rs
@@ -25,7 +25,7 @@ use file_parsing::read_file_lines;
 
 use crate::kmer_counter::KmerCounter;
 
-/// Extract sequences that contain some kmers 
+/// Extract sequences that contain some kmers
 ///
 /// Also output the kmers that occur in the reads with their number of occurrences.
 #[allow(clippy::too_many_arguments)]
@@ -125,7 +125,7 @@ pub fn back_to_sequences<T: KmerCounter>(
     Ok(())
 }
 
-/// Extract sequences that contain some kmers 
+/// Extract sequences that contain some kmers
 ///
 /// Also output the kmers that occur in the reads with their number of occurrences.
 /// Instead of using a single input file takes a txt file with the path
