### 2024-09-16 6de483b
Applying typos and suggestions https://github.com/openjournals/joss-reviews/issues/7066#issuecomment-2351885620
