+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Back to sequences: find the origin of $k$-mers'
+title: 'Back to sequences: Find the origin of $k$-mers'
 tags:
   - rust
   - kmer
@@ -38,13 +38,13 @@ A vast majority of bioinformatics tools dedicated to the treatment of
   dispose of objects of fixed size that can be easily manipulated and
   compared to each other. A drawback is that the link between each
   $k$-mer and the original set of sequences to which it belongs is lost.
-  Given the volume of data considered in this context, finding back this
+  Given the volume of data considered in this context, recovering this
   association is costly. In this work, we present "`back_to_sequences`",
   a simple tool designed to index a set of $k$-mers of interest and to
   stream a set of sequences, extracting those containing at least one of
   the indexed $k$-mer. In addition, the occurrences of $k$-mers in the
   sequences can be provided. Our results show that `back_to_sequences`
-  streams $\approx200$ short read per millisecond, allowing to search
+  streams $\approx200$ short reads per millisecond, allowing to search
   $k$-mers in hundreds of millions of reads in a matter of a few
   minutes.
 
@@ -62,24 +62,25 @@ unknown sequencing strand) were complex to handle with OLC while being
 easy to handle or simply solved with the dBG
 approach [@li2012comparison].
 
-Recall that in the dBG assembly approach, 1. all $k$-mers (words of
-length $k$) from a set of reads are counted; 2. those with an abundance
-lower than a threshold are considered to contain sequencing errors
-and are discarded; 3. the remaining $k$-mers are organized in a dBG; 4.
-the paths of the dBG form the basis of the assembly, later improved
+Recall that in the dBG assembly approach, 
+1. All $k$-mers (words oflength $k$) from a set of reads are counted; 
+2. Those with an abundance lower than a threshold are considered to contain sequencing errors
+and are discarded; 
+3. The remaining $k$-mers are organized in a dBG; 
+4. The paths of the dBG form the basis of the assembly, later improved
 thanks to scaffolding tools [@huson2002greedy] such as the tools provided,
 for instance, by the Spades assembler [@bankevich2012spades].
 
 The usefulness of $k$-mers did not end with their use in dBGs. A large
 and redundant set of sequences, such as a sequencing read set, can be
-summarized by its set of $k$-mers. Among multiple fundamental tasks, it
+summarized by its set of $k$-mers. Among multiple fundamental tasks, this
 has been the basis for metagenome comparisons [@benoit2016multiple], for
 taxonomy characterization [@wood2019improved], for indexing
 purposes [@cracco2023extremely; @lemane2022kmtricks], for
 genotyping [@grytten2022kage], for species
 identification [@sarmashghi2019skmer], for transcript expression
 estimation [@zhang2014rna], or for variant
-discovery [@uricaru2015reference] to cite only a few.
+discovery [@uricaru2015reference] to cite only a few examples.
 
 One of the keys to the success of the use of $k$-mers is its low
 resource needs. Whatever the sequencing coverage, once
@@ -98,25 +99,25 @@ associated read identifier. This is not acceptable.
 
 There are many situations in which finding the origin of $k$-mers from a
 set in a set of reads is informative. For instance, this is the case in
-studies in which the output is composed of $k$-mers associated to
+studies in which the output is composed of $k$-mers associated with
 biological knowledge such as biological
 variants [@uricaru2015reference], or $k$-mers specific to a phenotypic
-trait [@lemane2022k] to cite a few. This approach can also be used for
+trait [@lemane2022k]. This approach can also be used for
 quality control [@plaza2015quality] or contamination
 removal [@gonzalez2023akmerbroom] for instance.
 
-Finding back the link between a $k$-mer and each original read in which
[…]
