+++ paper/paper.bib
@@ -69,7 +69,7 @@ doi={10.7717/peerj-cs.94},
   publisher={PeerJ Inc.}
 }
 @article{li2012comparison,
-  title={Comparison of the two major classes of assembly algorithms: overlap--layout--consensus and de-bruijn-graph},
+  title={Comparison of the two major classes of assembly algorithms: overlap--layout--consensus and de-Bruijn-graph},
   author={Li, Zhenyu and Chen, Yanxiang and Mu, Desheng and Yuan, Jianying and Shi, Yujian and Zhang, Hao and Gan, Jun and Li, Nan and Hu, Xuesong and Liu, Binghang and others},
   journal={Briefings in functional genomics},
   volume={11},
@@ -117,7 +117,7 @@ doi={10.1093/nar/gku1187},
 }
 
 @article{lemane2022kmtricks,
-  title={Kmtricks: efficient and flexible construction of bloom filters for large sequencing data collections},
+  title={Kmtricks: efficient and flexible construction of Bloom filters for large sequencing data collections},
   author={Lemane, T{\'e}o and Medvedev, Paul and Chikhi, Rayan and Peterlongo, Pierre},
   journal={Bioinformatics Advances},
   volume={2},
