### 2024-09-16 #7066 @pierrepeterlongo
Hi @majensen

Thank you for your careful reading and suggestions.

I applied all of them at the exception of:

- l.16: “occurences” -> distribution?
  - I proposed "occurrence positions".
- l.79 “one k-mer” -> “one k-mer (length 31)” (correct?)
  - that's correct, I indicated “one k-mer (with k=31)”

I also updated the https://b2s-doc.readthedocs.io/en/latest/results.html page adding notes about the section `Possible Alternatives`.

I'm waiting for your feedback before to create zenodo repositories.

Thanks again,
Pierre
