### 2024-09-18 #7066 @majensen
Thanks @pierrepeterlongo - I think it reads smoothly. Are you sure you don't want to capitalize "Bloom filters" in the refs (since Bloom is the name of the guy who invented them?). In any case, please go ahead and create the archive, and report the DOI back in this thread.
