+++ paper/paper.bib
@@ -117,7 +117,7 @@ doi={10.1093/nar/gku1187},
 }
 
 @article{lemane2022kmtricks,
-  title={Kmtricks: efficient and flexible construction of Bloom filters for large sequencing data collections},
+  title={TESTKmtricks: efficient and flexible construction of Bloom filters for large sequencing data collections},
   author={Lemane, T{\'e}o and Medvedev, Paul and Chikhi, Rayan and Peterlongo, Pierre},
   journal={Bioinformatics Advances},
   volume={2},
