+++ paper/paper.bib
@@ -117,7 +117,7 @@ doi={10.1093/nar/gku1187},
 }
 
 @article{lemane2022kmtricks,
-  title={Kmtricks: efficient and flexible construction of BBloom filters for large sequencing data collections},
+  title={Kmtricks: efficient and flexible construction of Bloom filters for large sequencing data collections},
   author={Lemane, T{\'e}o and Medvedev, Paul and Chikhi, Rayan and Peterlongo, Pierre},
   journal={Bioinformatics Advances},
   volume={2},
@@ -227,7 +227,7 @@ doi={10.1186/s12864-015-1406-7},
 }
 
 @article{gonzalez2023akmerbroom,
-  title={aKmerBroom: Ancient oral DNA decontamination using Bloom filters on k-mer sets},
+  title={aKmerBroom: Ancient oral DNA decontamination using BBloom filters on k-mer sets},
   author={Gonz{\'a}lez, Camila Duitama and Rangavittal, Samarth and Vicedomini, Riccardo and Chikhi, Rayan and Richard, Hugues},
   journal={Iscience},
   volume={26},
