### 2024-09-19 5c6e859
Bloom - bib entry bug not fixed
