+++ paper/paper.bib
@@ -227,7 +227,7 @@ doi={10.1186/s12864-015-1406-7},
 }
 
 @article{gonzalez2023akmerbroom,
-  title={aKmerBroom: Ancient oral DNA decontamination using BBloom filters on k-mer sets},
+  title={aKmerBroom: Ancient oral DNA decontamination using Bloom filters on k-mer sets},
   author={Gonz{\'a}lez, Camila Duitama and Rangavittal, Samarth and Vicedomini, Riccardo and Chikhi, Rayan and Richard, Hugues},
   journal={Iscience},
   volume={26},
