### 2024-09-19 #7066 @pierrepeterlongo
Hi @majensen

That's crazy but I can't change 'b' to 'B' in the references.
If I change "bloom" from "Bloom" in the .bib, this has no effect on the pdf, but if I change "bloom" from "BBloom" for instance, with two 'B's the pdf changes accordingly.

Anyway I created the two zenodo repositories:
- https://doi.org/10.5281/zenodo.13794495 containing the documentation as a pdf
- https://doi.org/10.5281/zenodo.13794732 containing the soft repository, and linked to the previous zeonodo repo for the documentation, as you suggested https://github.com/openjournals/joss-reviews/issues/7066#issuecomment-2331388851

Best,
Pierre
