### 2024-09-19 #7066 @majensen
One minor update @pierrepeterlongo - can you make the title of the repo https://doi.org/10.5281/zenodo.13794732 the same as that of the paper, ie "Back to sequences: find the origin of k-mers"? This is something JOSS requires.
