### 2024-09-20 #7066 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
