### 2024-09-20 #7066 @majensen
@pierrepeterlongo  - I should explain that I as the topic editor do the "recommend-accept" command, which does some checks and then informs the Editors-in-chief. One of them will come soon and do a final check - they may have questions or minor fixes. When they are satisfied, they'll push the big button that publishes the paper officially.
Thanks!
