### 2024-09-23 #7066 @Kevin-Mattheus-Moerman
@pierrepeterlongo as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I will now process some final checks:

- [ ] Software license tag listed here matches a tagged release

- [ ]
