### 2024-09-23 #7066 @pierrepeterlongo
Just a last comment, if it is possible, maybe remove the '$'s  in the title on the JOSS page:
