### 2024-09-26 #7066 @Kevin-Mattheus-Moerman
@openjournals/dev @arfon how do we go about removing those dollar symbols? :point_up:
