### 2024-09-30 #7066 @Kevin-Mattheus-Moerman
Great, since that is all set we are good to close this review.

Congratulations @pierrepeterlongo on this JOSS publication!

Thanks @majensen for editing!
